requirements.txt
django
psycopg2-binary
cohere
sentence-transformers
python-dotenv
pgvector
pypdf
django-extensions
Werkzeug
django-allauth[socialaccount]
openai
django-debug-toolbar
mypy
django-stubs
anthropic
anthropic[bedrock]
tenacity
uvicorn
whitenoise
typeguard
pydantic
channels
daphne
google-generativeai
django-storages[google]
django-storages[s3]
pytest
pytest-cov
pytest-django
boto3
websockets
debugpy
channels-redis
django-stubs
celery[redis]
pdf2image
pillow
google-generativeai
dotenv
pytesseract
google-genai
chardet
tiktoken
trafilatura
beautifulsoup4
selenium>=4.0.0
webdriver-manager>=4.0.0
astropy
numpy
requests-oauthlib